requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
xgboost
shap
google-generativeai
python-dotenv
fairlearn
shap
matplotlib
